+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## next
 
+* Add a `Hashable` instance for `Fin`.
+
 ## 2.3.0.0 -- 2026-03-06
 
   * BREAKING: Remove the `Iso`s `fin0Void`, `fin1Unit`, and `fin2Bool`.